refactor: extract symbol logic into domain modules and improve role resolver
Split ansibleSymbols.ts into focused modules (handlerSymbols, variableSymbols,
moduleSymbols, symbolDefinitions). Export shared patterns from ansible.ts,
add workspace symbol cache invalidation, fix nested /roles/ path resolution,
and add escapeRegex utility.
