+++ packages/ansible-language-server/src/providers/completionProvider.ts
@@ -256,7 +256,7 @@ export async function doCompletion(
         isCursorInsideJinjaBrackets(document, position, path)
       ) {
         const rolesPaths = (await context.ansibleConfig).roles_paths;
-        const varCompletion: CompletionItem[] = await getVarsCompletion(
+        const varCompletion: CompletionItem[] = getVarsCompletion(
           document.uri,
           path,
           rolesPaths,
